refactor(builder): extract session-detail helpers to fix Fast Refresh lint errors
